+++ src/Framebuffer.hpp
@@ -5,6 +5,6 @@ void setPixel(int x, int y);
 void setPixelOff(int x, int y);
 void setPixelOff(int x);
 void clearScreen();
-void initFramebuffer(float* wave_table);
-void table2Screen(float* wave_table);
+void initFramebuffer(double* wave_table);
+void table2Screen(double* wave_table);
 #endif
