+++ src/WaveOsc.hpp
@@ -1,7 +1,7 @@
 // WaveOsc.h
 #ifndef WAVEOSC_H
 #define WAVEOSC_H
-float getWavetableValue(float* wave_table);
-void setfreq(float f);
+double getWavetableValue(double* wave_table);
+void setfreq(double f);
 #endif
